Refactor APISettingsActivity and ListAdapter
